CLN: misplaced indexing tests
