+++ pandas/tests/indexes/datetimes/test_partial_slicing.py
@@ -1,7 +1,6 @@
 """ test partial slicing on Series/Frame """
 
 from datetime import datetime
-import operator
 
 import numpy as np
 import pytest
@@ -411,40 +410,6 @@ class TestSlicing:
 
         result = df.loc["2016-10-01T00:00:00":]
         tm.assert_frame_equal(result, df)
-
-    @pytest.mark.parametrize(
-        "datetimelike",
-        [
-            Timestamp("20130101"),
-            datetime(2013, 1, 1),
-            np.datetime64("2013-01-01T00:00", "ns"),
-        ],
-    )
-    @pytest.mark.parametrize(
-        "op,expected",
-        [
-            (operator.lt, [True, False, False, False]),
-            (operator.le, [True, True, False, False]),
-            (operator.eq, [False, True, False, False]),
-            (operator.gt, [False, False, False, True]),
-        ],
-    )
-    def test_selection_by_datetimelike(self, datetimelike, op, expected):
-        # GH issue #17965, test for ability to compare datetime64[ns] columns
-        # to datetimelike
-        df = DataFrame(
-            {
-                "A": [
-                    Timestamp("20120101"),
-                    Timestamp("20130101"),
-                    np.nan,
-                    Timestamp("20130103"),
-                ]
-            }
-        )
-        result = op(df.A, datetimelike)
-        expected = Series(expected, name="A")
-        tm.assert_series_equal(result, expected)
 
     @pytest.mark.parametrize(
         "start",
