+++ pandas/tests/indexes/period/test_indexing.py
@@ -205,6 +205,7 @@ class TestGetItem:
                 # GH7116
                 # these show deprecations as we are trying
                 # to slice with non-integer indexers
+                # FIXME: don't leave commented-out
                 # with pytest.raises(IndexError):
                 #    idx[v]
                 continue
@@ -813,12 +814,6 @@ class TestGetValue:
         with tm.assert_produces_warning(FutureWarning):
             result2 = idx2.get_value(input2, p1)
         tm.assert_series_equal(result2, expected2)
-
-    def test_loc_str(self):
-        # https://github.com/pandas-dev/pandas/issues/33964
-        index = period_range(start="2000", periods=20, freq="B")
-        series = Series(range(20), index=index)
-        assert series.loc["2000-01-14"] == 9
 
     @pytest.mark.parametrize("freq", ["H", "D"])
     def test_get_value_datetime_hourly(self, freq):
