+++ pandas/tests/series/test_api.py
@@ -182,3 +182,12 @@ class TestSeriesMisc:
         ser = Series(dtype=object)
         with tm.assert_produces_warning(None):
             inspect.getmembers(ser)
+
+    def test_unknown_attribute(self):
+        # GH#9680
+        tdi = pd.timedelta_range(start=0, periods=10, freq="1s")
+        ser = Series(np.random.normal(size=10), index=tdi)
+        assert "foo" not in ser.__dict__.keys()
+        msg = "'Series' object has no attribute 'foo'"
+        with pytest.raises(AttributeError, match=msg):
+            ser.foo
